requirements.txt
Babel==2.8.0
blinker==1.4
certifi==2020.4.5.1
click==7.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
passlib==1.7.2
psycopg2==2.8.5
python-dotenv==0.14.0
pytz==2020.1
speaklater==1.3
SQLAlchemy==1.3.17
Werkzeug==1.0.0
wincertstore==0.2
WTForms==2.3.1
gunicorn==20.0.4